@@ -3,36 +3,36 @@
 	"version": "0.1.0",
 	"private": true,
 	"dependencies": {
-		"@babel/core": "^7.18.10",
+		"@babel/core": "^7.20.12",
 		"@babel/plugin-syntax-jsx": "^7.18.6",
-		"@emotion/babel-plugin": "^11.10.0",
-		"@emotion/react": "^11.10.0",
-		"@mantine/core": "^5.1.6",
-		"@mantine/dropzone": "^5.1.6",
-		"@mantine/form": "^5.1.6",
-		"@mantine/hooks": "^5.1.6",
-		"@mantine/notifications": "^5.1.6",
-		"@mantine/prism": "^5.1.6",
-		"@testing-library/dom": "^8.17.1",
+		"@emotion/babel-plugin": "^11.10.5",
+		"@emotion/react": "^11.10.5",
+		"@mantine/core": "^5.10.0",
+		"@mantine/dropzone": "^5.10.0",
+		"@mantine/form": "^5.10.0",
+		"@mantine/hooks": "^5.10.0",
+		"@mantine/notifications": "^5.10.0",
+		"@mantine/prism": "^5.10.0",
+		"@testing-library/dom": "^8.19.1",
 		"@testing-library/jest-dom": "^5.16.5",
-		"@testing-library/react": "^13.3.0",
+		"@testing-library/react": "^13.4.0",
 		"@testing-library/user-event": "^14.4.3",
-		"@types/jest": "^28.1.7",
-		"@types/node": "^18.7.5",
-		"@types/react": "^18.0.17",
-		"@types/react-dom": "^18.0.6",
-		"@vitejs/plugin-react": "^2.0.1",
-		"autoprefixer": "^10.4.8",
-		"immer": "^9.0.15",
-		"postcss": "^8.4.16",
+		"@types/jest": "^28.1.8",
+		"@types/node": "^18.11.18",
+		"@types/react": "^18.0.26",
+		"@types/react-dom": "^18.0.10",
+		"@vitejs/plugin-react": "^2.2.0",
+		"autoprefixer": "^10.4.13",
+		"immer": "^9.0.17",
+		"postcss": "^8.4.21",
 		"react": "^18.2.0",
 		"react-dom": "^18.2.0",
-		"react-router-dom": "6",
-		"tabler-icons-react": "^1.54.0",
-		"typescript": "^4.7.4",
-		"usehooks-ts": "^2.6.0",
+		"react-router-dom": "^6.6.1",
+		"tabler-icons-react": "^1.56.0",
+		"typescript": "^4.9.4",
+		"usehooks-ts": "^2.9.1",
 		"web-vitals": "^2.1.4",
-		"zustand": "^4.0.0"
+		"zustand": "^4.2.0"
 	},
 	"scripts": {
 		"dev": "vite",
@@ -58,7 +58,7 @@
 		]
 	},
 	"devDependencies": {
-		"vite": "^3.0.7",
-		"vite-plugin-svgr": "^2.2.1"
+		"vite": "^3.2.5",
+		"vite-plugin-svgr": "^2.4.0"
 	}
 }
